Update templates to auto add Test action to Scheme
We discovered that Xcode will add a test bundle to the Test action for
your project's scheme iff the template identifer matches this ridiculous
com.apple identifier.

Additionally this seems to happen asynchronously, so we need to sleep in
the tests for a second in order for this to happen somewhat reliably.